Recursive solution typo
Good article. Think there's a mistake in the recursive solution though (see diff). When the `*` is at the end of the current `pattern`, taking the whole pattern and seeing if the str is at the start isn't the right check (abcd is not at the start of ab*). Copying the code and running it against the third example case confirms this.